New CSS file
My aim was to left align the text for both pages by creating a new css page for this pages style and then left-aligning it.

I Made a new CSS file for this pages style. I duplicated the style from home.module, but just left aligned the text, so everything else should be the same.

🤞this works.